requirements.txt
# --- Data Pipeline ---
yfinance>=0.2.31
fredapi>=0.5.1
pandas>=2.0
numpy>=1.24
pyarrow>=14.0

# --- Signals & ML ---
ta>=0.11.0
lightgbm>=4.0
xgboost>=2.0
scikit-learn>=1.3
shap>=0.44
optuna>=3.4
scipy>=1.11

# --- Backtesting & Reporting ---
matplotlib>=3.7
seaborn>=0.13

# --- Backend API ---
fastapi>=0.109
uvicorn>=0.27
sqlalchemy>=2.0
apscheduler>=3.10
aioapns>=3.0

# --- Config ---
pyyaml>=6.0
